implemented menu and basic state machine, gave the player movement and an image,

+++ trunk/ZombieGentlemen_SeniorProject/weapon.cpp
@@ -24,7 +24,7 @@ int Weapon::weaponDamageType(void)
 {
 	return 0;
 }
-float Weapon::cost(int cost)
+float Weapon::cost(float cost)
 {
 	
 	return cost;
